requirements.txt
azure-mgmt-resource>=1.18.0
azure-identity>=1.5.0